fix(character-extraction): lift ReferenceImageViewer to CharacterList level
Viewer was absolute inset-0 inside ModalShell content area, constrained
to the modal's natural content height. Moving it to CharacterList gives
it the full step area (h-full) to fill at z-30.

- CharacterList owns viewingImageId state and deleteReferenceImage
- CharacterModal accepts onImageClick prop, no longer owns viewer state
- ModalShell content wrapper reverts to relative flex flex-col
