+++ frontend/src/features/scene-engine/steps/character-extraction/components/CharacterModal.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import { ModalShell } from "../../../components/ModalShell";
 import PromptInput from "../../../components/PromptInput";
 import { useSceneEngine } from "../../../context";
@@ -6,50 +5,22 @@ import { CharacterAttributes } from "../CharacterAttributes";
 import type { CharacterBundle } from "../character-extraction.types";
 import { useCharacterExtraction } from "../hooks/useCharacterExtraction";
 import { RefImageTray } from "./RefImageTray";
-import { ReferenceImageViewer } from "./ReferenceImageViewer";
 
 type CharacterModalProps = {
   bundle: CharacterBundle;
   onDismiss: () => void;
+  onImageClick: (imageId: string) => void;
 };
 
-export function CharacterModal({ bundle, onDismiss }: CharacterModalProps) {
+export function CharacterModal({ bundle, onDismiss, onImageClick }: CharacterModalProps) {
   const { projectId, storyId } = useSceneEngine();
-  const {
-    refineCharacter,
-    isRefining,
-    uploadReferenceImage,
-    isUploading,
-    deleteReferenceImage,
-    isDeleting,
-  } = useCharacterExtraction(projectId, storyId);
+  const { refineCharacter, isRefining, uploadReferenceImage, isUploading } =
+    useCharacterExtraction(projectId, storyId);
 
-  const [viewingImageId, setViewingImageId] = useState<string | null>(null);
   const characterId = bundle.character.id;
-
-  const viewingImage = viewingImageId
-    ? bundle.referenceImages.find((r) => r.id === viewingImageId)
-    : null;
-
-  const handleDelete = () => {
-    if (!viewingImageId) return;
-    deleteReferenceImage(
-      { characterId, imageId: viewingImageId },
-      { onSuccess: () => setViewingImageId(null) },
-    );
-  };
 
   return (
     <ModalShell header={bundle.character.name} onDismiss={onDismiss}>
-      {viewingImage && (
-        <ReferenceImageViewer
-          img={viewingImage}
-          onBack={() => setViewingImageId(null)}
-          onDelete={handleDelete}
-          isDeleting={isDeleting}
-        />
-      )}
-
       <div className="flex min-h-0 flex-1">
         <div className="min-h-0 flex-1 overflow-y-auto p-3">
           <CharacterAttributes character={bundle.character} />
@@ -57,7 +28,7 @@ export function CharacterModal({ bundle, onDismiss }: CharacterModalProps) {
         <RefImageTray
           images={bundle.referenceImages}
           variant="modal"
-          onImageClick={setViewingImageId}
+          onImageClick={onImageClick}
         />
       </div>
       <div className="shrink-0 border-t border-border">
